BondStateMachine: Allow skipping confirm for some remotes
When pairing two of the Nvidia Shield accessories, a popup would show up
stating that the accessory was an incoming pairing request and needs to
be accepted. The official Nvidia firmware has a whitelist of remotes
that skip this confirmation if pairing request is marked as originating
from the Android device. This change takes a similar approach, but in a
more flexible manner. The main intent is to allow these accessories to
be paired via the pairing intent, which needs to complete with no user
interaction. Previously, the popup would prevent this from succeeding.

Change-Id: Ib5a0226858f5745a20e4cd166500aecdcf1f3354
Signed-off-by: S I P U N <sipunkumar85@gmail.com>
Signed-off-by: NRanjan-17 <nalinishranjan05@gmail.com>
Signed-off-by: Mrick343 <mrinmoy8429de@gmail.com>